pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "trs-repro"
version = "0.1.0"
description = "End-to-end reproduction of Thinking with Reasoning Skills"
readme = "README.md"
requires-python = ">=3.10"
authors = [{name = "OpenAI"}]
dependencies = [
  "datasets>=2.20.0",
  "faiss-cpu>=1.8.0",
  "litellm>=1.61.0",
  "orjson>=3.10.0",
  "pydantic>=2.7.0",
  "PyYAML>=6.0.1",
  "rank-bm25>=0.2.2",
  "sentence-transformers>=3.0.0",
  "sympy>=1.13.0",
  "tenacity>=8.3.0",
  "tqdm>=4.66.0",
  "typer>=0.12.0",
  "python-dotenv>=1.0.1",
  "numpy>=1.26.0",
  "pandas>=2.2.0"
]

[project.optional-dependencies]
math = [
  "math-verify[antlr4_11_0]>=0.7.0"
]

[project.scripts]
trs = "trs.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
